+++ js/photo.js
@@ -1,0 +1,13 @@
+const createPhoto = (actions) => {
+  const {getId, getUrl, getDescription, getLikes, getComments} = actions;
+
+  const id = getId();
+  const url = getUrl(id);
+  const description = getDescription(id);
+  const likes = getLikes();
+  const comments = getComments();
+
+  return {id, url, description, likes, comments};
+};
+
+export {createPhoto};
